refactor(package): MVC 레이어에 맞춰 패키지 구조 정렬
