+++ src/main/java/lotto/model/domain/Lotto.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.model.domain;
 
 import java.util.Collections;
 import java.util.HashSet;
