+++ src/main/java/lotto/model/domain/WinningCriteria.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.model.domain;
 
 public enum WinningCriteria {
     FIRST(6, false, 2_000_000_000L),
