+++ src/main/java/lotto/view/OutputView.java
@@ -1,10 +1,10 @@
-package lotto;
+package lotto.view;
 
 import java.text.DecimalFormat;
 import java.util.List;
 import java.util.Map;
-import lotto.domain.Lotto;
-import lotto.domain.WinningCriteria;
+import lotto.model.domain.Lotto;
+import lotto.model.domain.WinningCriteria;
 
 public class OutputView {
     public void printError(String error) {
